analytics/requirements.txt
# MongoDB Connection
pymongo>=4.6.0
python-dotenv>=1.0.0

# Data Analysis
pandas>=2.1.0
numpy>=1.24.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Geospatial Analysis
geopy>=2.4.0

# Statistical Analysis
statsmodels>=0.14.0

# Time Series Analysis
prophet>=1.1.5

# Reporting
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# API Server
flask>=3.0.0
flask-cors>=4.0.0

# API Integration (if needed)
requests>=2.31.0
